Pendulum created (needs debugging)
