+++ Assets/Scripts/Obstacles/Pendulum.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 123c5b7745ca68647ad25b373dd91e99
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
